pyproject.toml
[project]
name = "daily-agenda-email"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "apscheduler>=3.10,<4",
    "google-auth-oauthlib",
    "google-api-python-client",
    "azure-identity",
    "azure-keyvault-secrets",
    "azure-communication-email",
    "httpx",
    "pyyaml",
    "jinja2",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-watch",
    "httpx",
    "ruff",
    "pyright",
    "pre-commit",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
venvPath = "/home/vscode"
venv = ".venv"
pythonVersion = "3.12"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
